fix: one component per file - Badge edition (#443)
