+++ src/index.js
@@ -1,9 +1,9 @@
 export { default as ActionBar } from './ActionBar/ActionBar';
 export { Alert } from './Alert/Alert';
-export { Badge } from './Badge/Badge';
-export { Label } from './Badge/Badge';
-export { Status } from './Badge/Badge';
-export { Counter } from './Badge/Badge';
+export { default as Badge } from './Badge/Badge';
+export { default as Label } from './Badge/Label';
+export { default as Status } from './Badge/Status';
+export { default as Counter } from './Badge/Counter';
 export { Breadcrumb } from './Breadcrumb/Breadcrumb';
 export { BreadcrumbItem } from './Breadcrumb/Breadcrumb';
 export { default as Button } from './Button/Button';
